Ollama and Executable updates
Made it so that ollama will attempt to run on the first message. An error message will be provided to the user with instructions on how to install ollama and llava if it fails to run, and will tell the user to try their prompt again if it is able to run. This works for the newly created executable as well. Also changed the name of the executable form run_Bob to BobAI.